@@ -1,13 +1,16 @@
 module github.com/fixme_my_friend/hw05_parallel_execution
 
-go 1.16
+go 1.19
+
+require (
+	github.com/stretchr/testify v1.7.0
+	go.uber.org/goleak v1.1.10
+)
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/stretchr/testify v1.7.0
-	go.uber.org/goleak v1.1.10
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	golang.org/x/lint v0.0.0-20201208152925-83fdc39ff7b5 // indirect
 	golang.org/x/tools v0.1.0 // indirect
-	gopkg.in/yaml.v2 v2.4.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )
