@@ -1,10 +1,11 @@
 module github.com/fixme_my_friend/hw06_pipeline_execution
 
-go 1.16
+go 1.19
+
+require github.com/stretchr/testify v1.7.0
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/stretchr/testify v1.7.0
-	gopkg.in/yaml.v2 v2.4.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )
